Fix to handle case of more than 4 arguments
git-svn-id: svn://svn.code.sf.net/p/nuttx/code/trunk@1945 42af7a65-404d-4744-a932-0658087f49c3

+++ misc/buildroot/toolchain/nxflat/ldnxflat.c
@@ -960,7 +960,7 @@ resolve_segment_relocs(bfd *input_bfd, segment_info *inf, asymbol **syms)
                     exit(1);
                   }
 
-                vdbg("  Modified opcode: %08lx\n", temp);
+                vdbg("  Modified opcode: %08lx\n", (long)temp);
 #ifdef ARCH_BIG_ENDIAN
                 *target = (long)nxflat_swap32(temp);
 #else
@@ -1028,7 +1028,7 @@ resolve_segment_relocs(bfd *input_bfd, segment_info *inf, asymbol **syms)
 
                 temp |= saved & (~how_to->dst_mask);
 
-                vdbg("  Modified opcode: %08lx\n", temp);
+                vdbg("  Modified opcode: %08lx\n", (long)temp);
 #ifdef ARCH_BIG_ENDIAN
                 *target = (long)nxflat_swap32(temp);
 #else
